Add files via upload
This is a simple Rock, Paper, Scissors game. A player competes against a randomly generated computer response.
logically speaking, there should be a reset button to start all over again, and maybe a scoreboard.
Still need to work on the input pass to the simple logic function and then show the result to the player.
no CSS due to time constraints (2 Hours).